requirements.txt
aiogram==3.13.1
aiohttp==3.10.10
python-calamine==0.2.3
apscheduler==3.10.4
python-dotenv==1.0.1
